+++ devito/petsc/iet/passes.py
@@ -18,12 +18,12 @@ from devito.petsc.types import (
 from devito.petsc.types.macros import petsc_func_begin_user
 from devito.petsc.iet.nodes import PetscMetaData, petsc_call
 from devito.petsc.internals import core_metadata, petsc_languages
-from devito.petsc.iet.callback_builder import (
+from devito.petsc.iet.callbacks import (
     BaseCallback, CoupledCallback, populate_matrix_context, get_user_struct_fields
 )
-from devito.petsc.iet.object_builder import BaseObjectBuilder, CoupledObjectBuilder, objs
-from devito.petsc.iet.pre_solver import BaseSetup, CoupledSetup, make_core_petsc_calls
-from devito.petsc.iet.run_solver import Solver, CoupledSolver
+from devito.petsc.iet.objects import BaseObjectBuilder, CoupledObjectBuilder, objs
+from devito.petsc.iet.builder import BaseSetup, CoupledSetup, make_core_petsc_calls
+from devito.petsc.iet.solver import Solver, CoupledSolver
 from devito.petsc.iet.time_dependence import TimeDependent, TimeIndependent
 from devito.petsc.iet.logging import PetscLogger
